pyproject.toml
[tool.poetry]
name = "traintogain"
version = "0.1.0"
description = ""
authors = ["Damian Bednarek <damian.bednarekk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
django-configurations = "^2.5.1"
load-dotenv = "^0.1.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
faker = "^26.0.0"
factory-boy = "^3.3.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
